% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{create_AC_Product_by_Link}
\alias{create_AC_Product_by_Link}
\title{Create an AC_PRODUCT table consisting of aries cases linked to the list of speicifed enertia codes.}
\usage{
create_AC_Product_by_Link(
  enertia_codes,
  server = "extsql01",
  database = "Aries",
  user = Sys.getenv("USERNAME"),
  overwrite = "Y",
  export = FALSE
)
}
\arguments{
\item{enertia_codes}{a character vector.}

\item{server}{a character vector.}

\item{database}{a character vector.}

\item{user}{a character vector.}

\item{overwrite}{a character vector.}

\item{export}{a boolean value}
}
\value{
A dataframe containing production data formated for the AC_PRODUCT table in Aries
}
\description{
The following functin will make a connection to the Enertia database located on the Enertia01B Server.
It will then take the user defined 'enertia_codes' vector which contains one or more enertia_codes to query.
The function will then return a dataframe object which contains aries cases linked to the specified enertia codes and the associated production data.
the data frame is returned in the same format as the Aries AC_PRODUCT table.
the export variable can be set equal to TRUE to export the output to the users desktop.
}
\examples{
create_AC_Product_by_Link(enertia_codes = c("49.1018.0010.00", "49.1018.0210.00"), user = "user01", overwrite = "N", export = TRUE)
}
