% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_Aries_Enertia_Links_by_Area}
\alias{get_Aries_Enertia_Links_by_Area}
\title{Retrieve Aries-Enertia links as a data frame for a user defined vector of asset team codes.}
\usage{
get_Aries_Enertia_Links_by_Area(asset_team_code, export = FALSE)
}
\arguments{
\item{asset_team_code}{a character vector.}

\item{export}{a boolean value}
}
\value{
A dataframe containing the Aries-Enertia links for a specified vector of asset team codes
}
\description{
The following functin will make a connection to the Enertia database located on the Enertia01B Server.
It will then take the user defined 'asset_team_code' vector which contains one or more asset teams to query.
The function will then return a dataframe object with all the current aries_enertia links that exists at the time of the query.
the export variable can be set equal to TRUE to export the output to the users desktop.
}
\examples{
get_Aries_Enertia_Links_by_Area("RMR", export = TRUE)
}
