% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_Aries_Enertia_Links_by_Enertia_Code}
\alias{get_Aries_Enertia_Links_by_Enertia_Code}
\title{Retrieve Aries-Enertia links as a data frame for a user defined vector of enertia codes.}
\usage{
get_Aries_Enertia_Links_by_Enertia_Code(enertia_codes, export = FALSE)
}
\arguments{
\item{enertia_codes}{a character vector.}

\item{export}{a boolean value}
}
\value{
A dataframe containing the Aries-Enertia links for a specified vector of enertia codes
}
\description{
The following functin will make a connection to the Enertia database located on the Enertia01B Server.
It will then take the user defined 'enertia_codes' vector which contains one or more enertia codes to query.
The function will then return a dataframe object with all the current aries_enertia links that exists at the time of the query.
the export variable can be set equal to TRUE to export the output to the users desktop.
}
\examples{
get_Aries_Enertia_Links_by_Enertia_Code("49.1018.0010.00", export = TRUE)
}
