% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_Daily_Enertia_Production}
\alias{get_Daily_Enertia_Production}
\title{Retrieve Daily Enertia Production Data.}
\usage{
get_Daily_Enertia_Production(well_comp_codes, months, export = FALSE)
}
\arguments{
\item{well_comp_codes}{A string.}

\item{months}{An integer}

\item{export}{A bolean.}
}
\value{
A dataframe containing Daily Production data for a specified number of months since the current date
}
\description{
The following functin will make a connection to the ofm database located on the Enertia01B Server.
It will then take the well_comp_codes vector which contains a sequence of Enertia completion codes to query.
The function will then return a dataframe object with all the enertia daily production data that exists for each
Enertia completion code within the well_comp_codes vector for the specified number of months since the current date.
}
\examples{
get_Enertia_Production("49.1016.0022.00", 5)
}
