% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_dbhandle}
\alias{get_dbhandle}
\title{Create a SQL Connection}
\usage{
get_dbhandle(server, database)
}
\arguments{
\item{server}{A string.}

\item{database}{A string.}
}
\value{
The handle containing the database connection
}
\description{
This function opens a connection to the specified Server and Database parameters passed to it.
}
\examples{
get_dbhandle("extsql01", "Aries")
}
