% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_Economics}
\alias{get_Economics}
\title{Retrieve Aries economic data.}
\usage{
get_Economics(server, database, PROPNUM, scenario_name,
  view_by = "Scenario")
}
\arguments{
\item{server}{String containing the server name where the Aries database is located.}

\item{database}{String contiaing the name of the Aries database.}

\item{PROPNUM}{Sharacter vector containing a list of Aries PROPNUM to query from the database.}

\item{scenario_name}{String containing the name of the scenario for which economic lines should be queried.}

\item{view_by}{String containing the method as to which lines are to be filtered out (see description...).}
}
\value{
A list object containing a dataframe of economic data for each Aries economic section.
}
\description{
This function returns economic lines as a list by section based on the specified Data Source and Scenario Name.
PROPNUM is a vector of Aries PROPNUM to be queried.
The view_by parameter determines how the function queries the data:
   "All" = all lines,
   "Full Scenario" = All qualified lines within desired scenario,
   "Scenario" = Highest qualified line specific to scenario
}
\examples{
get_Economics("extsql01", "Aries", "T64EO3SCKW", "FINAL0119_WF", "Scenario")
}
