% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_Enertia_Lineage_by_API12}
\alias{get_Enertia_Lineage_by_API12}
\title{Retrieve Enertia Lineage as a data frame.}
\usage{
get_Enertia_Lineage_by_API12(API12, export = FALSE)
}
\arguments{
\item{API12}{A string.}

\item{export}{A bolean.}
}
\value{
A dataframe containing the Enertia Lineage for a specified vector of API12 Numbers
}
\description{
The following functin will make a connection to the ofm database located on the Enertia01B Server.
It will then take the API12 vector which contains a sequence of API12 Numbers to query.
The function will then return a dataframe object with all enertia entities that exists for each
API12 within the API12 vector.
}
\examples{
get_Enertia_Lineage_by_API("5002921693")
get_Enertia_Lineage_by_API(c("5002921693", "5002921905"))
}
