% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_Enertia_Lineage_by_Field_Code}
\alias{get_Enertia_Lineage_by_Field_Code}
\title{Retrieve Enertia Lineage as a data frame from ofm tables in enertia03.}
\usage{
get_Enertia_Lineage_by_Field_Code(field_code)
}
\arguments{
\item{field_code}{A string.}
}
\value{
A dataframe containing the Enertia Lineage of a set of Asset Team Codes
}
\description{
The following functin will make a default connection to the ofm tables in the enertia03 server/HEC_Repository database.
It will then take the field_code vector which contains a sequence of Enertia Field Codes to query.
The function will then return a dataframe object with all enertia entities that exists for each
Field Code within the field_code vector.
}
\examples{
get_Enertia_Lineage_by_Field_Code("49.1017")
get_Enertia_Lineage_by_Field_Code(c("49.1017", "49.1015")
}
