% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_Input_Settings}
\alias{get_Input_Settings}
\title{Queries Aries Input Settings}
\usage{
get_Input_Settings(user_input_setting, server, database, export = FALSE)
}
\arguments{
\item{user_input_setting}{String containing the name of the input settings in Aries.}

\item{server}{String containing the server name where the Aries database is located.}

\item{database}{String containing the Aries database name.}

\item{export}{A bolean.}
}
\value{
Returns a list object containing the As of Date, Base Date, PW table, common lines, and default lines for a specified user input settings.
}
\description{
This function will retrieve the As of Date, Base Date, PW table, common lines, and default lines for the input setting specified by the user.
}
\examples{
get_Input_Settings(user_input_settings = "HEC0119", server = "extsql01", database = "Aries")
}
