% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{QC_Aries_Enertia_Links}
\alias{QC_Aries_Enertia_Links}
\title{QC the existing Aries-Enertia links and lineage by returning a data frame for a user defined vector of asset team code.}
\usage{
QC_Aries_Enertia_Links(area, export = FALSE)
}
\arguments{
\item{area}{a character vector.}

\item{export}{a boolean value}
}
\value{
A dataframe containing QC flags for Aries-Enertia links and associated lineage
}
\description{
The following functin will make a connection to the Enertia database located on the Enertia01B Server.
It will then take the user defined 'area' vector which contains one or more asset team codes to query.
The function will then return a dataframe object which contains QC flags related to the existing aries-enertia links and the associated lineage.
#' the export variable can be set equal to TRUE to export the output to the users desktop.
}
\examples{
QC_Aries_Enertia_Links(c("ANS", "CIO"), export = TRUE)
}
