% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_Economics}
\alias{get_Economics}
\title{Retrieve Aries economic data.}
\usage{
get_Economics(
  server,
  database,
  PROPNUM,
  scenario_name,
  view_by = "Scenario",
  export = FALSE
)
}
\arguments{
\item{server}{A string.}

\item{database}{A string.}

\item{PROPNUM}{A String}

\item{scenario_name}{A string}

\item{view_by}{A string}

\item{export}{A bolean.}
}
\value{
A dataframe containing Economic data
}
\description{
This function returns economic lines as a list by section based on the specified Data Source and Scenario Name.
PROPNUM is a vector of PROPNUMs to be queried.
The View_by parameter determines how the function queries the data:
   "All" = all lines,
   "Full Scenario" = All qualified lines within desired scenario,
   "Scenario" = Highest qualified line specific to scenario
}
\examples{
get_Economics("extsql01", "Aries", "T64EO3SCKW", "FINAL0119_WF", "Scenario")
}
