% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_Enertia_Lineage_by_API}
\alias{get_Enertia_Lineage_by_API}
\title{Retrieve Enertia Lineage as a data frame.}
\usage{
get_Enertia_Lineage_by_API(API)
}
\arguments{
\item{API}{A string.}
}
\value{
A dataframe containing the Enertia Lineage for a specified vector of API Numbers
}
\description{
The following functin will make a default connection to the Enertia_Reports database located on the Default Enertia04 Server.
It will then take the API vector which contains a sequence of API Numbers to query.
The function will then return a dataframe object with all enertia entities that exists for each
API within the API vector.
}
\examples{
get_Enertia_Lineage_by_API("5002921693")
get_Enertia_Lineage_by_API(c("5002921693", "50029219050000"))
}
