% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reserves_Util.R
\name{get_Enertia_Lineage_by_Enertia_Code}
\alias{get_Enertia_Lineage_by_Enertia_Code}
\title{Retrieve Enertia Lineage as a data frame.}
\usage{
get_Enertia_Lineage_by_Enertia_Code(enertia_code)
}
\arguments{
\item{enertia_code}{A string.}
}
\value{
A dataframe containing the Enertia Lineage of a set of Asset Team Codes
}
\description{
The following functin will make a default connection to the Enertia_Reports database located on the Default Enertia04 Server.
It will then take the enertia_code vector which contains a sequence of Enertia Codes to query.
The function will then return a dataframe object with all enertia entities that exists for each
Enertia Code within the enertia_code vector.
}
\examples{
get_Enertia_Lineage_by_Enertia_Code("49.1017.0007.00")
get_Enertia_Lineage_by_Enertia_Code(c("49.1017.0007.00", "49.1017.0023.00"))
}
